// RedisCachePocClient/src/client/model/createMovieDto.ts
/**
 * MovieReviewer | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateMovieDto { 
    title: string;
    genre: string;
    director: string;
    plot: string;
}